+++ shared/domain/seasons/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/api/interactor/ObserveSeasonsInteractor.kt
@@ -28,7 +28,7 @@ fun SelectSeasonsByShowId.toSeasonsEntity(): SeasonUiModel {
         tvShowId = tv_show_id,
         name = name,
         overview = overview,
-        seasonNumber = season_number.toInt(),
+        seasonNumber = season_number,
         episodeCount = epiosode_count.toInt()
     )
 }
