+++ shared/remote/src/commonMain/kotlin/com/thomaskioko/tvmaniac/remote/api/TvShowsService.kt
@@ -18,7 +18,7 @@ interface TvShowsService {
 
     suspend fun getTvShowDetails(showId: Long): ShowDetailResponse
 
-    suspend fun getSeasonDetails(tvShowId: Long, seasonNumber: Int): SeasonResponse
+    suspend fun getSeasonDetails(tvShowId: Long, seasonNumber: Long): SeasonResponse
 
     suspend fun getTrendingShows(page: Int): TvShowsResponse
